Add Edit and Delete Actions
Commit adds Edit and Delete actions for source systems, as well as mappings. Select a source system or mapping from the table with the edit button to fire up the form modal with the editorial content. Delete a source system or mapping by clicking on the delete button in the table and selecting 'yes' in the confirm box.

Added:
- Edit method for Source Systems and Mappings
- Delete method for Source Systems and Mappings
- Confirm Box library

Modified:
- Updated and added the current harmonizes terms to the source JSON file
- Adapted the Types and Form Modal to the changes in the API