Add Insert/Patch/Delete Methods for Machine Annotation Services
Commit adds the Insert, Patch and Delete methods for Machine Annotation Services. Use the modal (apart file from Source Systems and Mappings) to add new machine annotation services. All values are free text at the moment, dependencies is treated as an array, and the Target Digital Object Filters is treated as an object based on the current harmonised terms of openDS.

Adds:
- New modal with form fields about the Machine Annotation Services
- Insert function for adding new MAS records
- Patch function for updating existing MAS records
- Delete function for removing existing MAS records

Modifies:
- Slight changes to the file hierarchy and names of folders/files to make things more general, and not all focused on Source Systems and Mappings like it was before.